getdehealth/requirements.txt
matplotlib
opencv-python
numpy
pillow
tkinter
ultralytics